set up route scheme and refactor WebViewScreen to use new WebViewer component
